+++ src/UI/Dashboard/Pages/WalletAccount.razor
@@ -14,7 +14,7 @@
 <br />
 
 <h3>Wallet Account</h3>
-<p>Information about <strong>@walletname\@accountname</strong></p>
+<p>Information about <strong>@walletname / @accountname</strong></p>
 
 @if (IndexService == null)
 {
