+++ src/UI/Dashboard/Pages/WalletReceive.razor
@@ -16,7 +16,7 @@
 
 <h3>Receive Coins</h3>
 
-<p>Generate receive address for <strong> @walletname\@accountname</strong></p>
+<p>Generate receive address for <strong> @walletname / @accountname</strong></p>
 
 @{
     var result = WalletService.WalletManager.GetUnusedAddress(new WalletAccountReference(walletname, accountname));
