+++ src/UI/Dashboard/Pages/WalletSend.razor
@@ -18,7 +18,7 @@
 
 <h3>Send Coins</h3>
 
-<p>Send coins from account <strong> @walletname\@accountname</strong></p>
+<p>Send coins from account <strong> @walletname / @accountname</strong></p>
 
 @{
     <div class="input-group mb-3">
